+++ src/frontend/stacked_windows/settings/security.py
@@ -20,14 +20,7 @@ class SecuritySettings(BaseSettings):
         )
         self.tmdb_api_key_entry = MaskedQLineEdit(self, masked=True)
 
-        tvdb_api_key_lbl = QLabel("TVDB Api Key", self)
-        tvdb_api_key_lbl.setToolTip(
-            "TVDB Api Key, required to determine file input metadata"
-        )
-        self.tvdb_api_key_entry = MaskedQLineEdit(self, masked=True)
-
         self.add_layout(create_form_layout(tmdb_api_key_lbl, self.tmdb_api_key_entry))
-        self.add_layout(create_form_layout(tvdb_api_key_lbl, self.tvdb_api_key_entry))
         self.add_layout(self.reset_layout, add_stretch=True)
 
         self._load_saved_settings()
@@ -37,14 +30,11 @@ class SecuritySettings(BaseSettings):
         """Applies user saved settings from the config"""
         payload = self.config.cfg_payload
         self.tmdb_api_key_entry.setText(payload.tmdb_api_key)
-        self.tvdb_api_key_entry.setText(payload.tvdb_api_key)
 
     @Slot()
     def _save_settings(self) -> None:
         self.config.cfg_payload.tmdb_api_key = self.tmdb_api_key_entry.text().strip()
-        self.config.cfg_payload.tvdb_api_key = self.tvdb_api_key_entry.text().strip()
         self.updated_settings_applied.emit()
 
     def apply_defaults(self) -> None:
         self.tmdb_api_key_entry.clear()
-        self.tvdb_api_key_entry.clear()